requirements.txt
fastapi==0.74.1
uvicorn==0.17.5
joblib==1.1.0
pyyaml==6.0
numpy==1.21.5
pandas==1.3.5
sklearn==0.0
mlflow==1.23.1
pydantic==1.9.0